requirements.txt
colorama==0.4.4
arcade==2.6.17
Pillow==9.5.0 